combine scattershot stuff under 1 header
